refactor(editor): extract App canvas geometry helpers
